feat(tests): updated tests for Long type

+++ app/src/test/kotlin/GraphTests.kt
@@ -11,23 +11,23 @@ import kotlin.test.assertEquals
 class GraphTests {
     @Test
     fun `undirected unweighted graph add vertex`() {
-        val graph = UndirectedUnweightedGraph<Int>()
+        val graph = UndirectedUnweightedGraph<Long>()
         graph.addVertex(Vertex(1, "1"))
         assertEquals(1, graph.vertices[1]?.id)
     }
 
     @Test
     fun `undirected unweighted graph get label`() {
-        val graph = UndirectedUnweightedGraph<Int>()
+        val graph = UndirectedUnweightedGraph<Long>()
         graph.addVertex(Vertex(1, "1"))
         assertEquals("1", graph.vertices[1]?.label)
     }
 
     @Test
     fun `undirected unweighted graph add edge`() {
-        val graph = UndirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
+        val graph = UndirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
         graph.addVertex(vertex1)
         graph.addVertex(vertex2)
         assertEquals(0, graph.addEdge(vertex1, vertex2))
@@ -35,9 +35,9 @@ class GraphTests {
 
     @Test
     fun `undirected weighted graph add edge`() {
-        val graph = UndirectedWeightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
+        val graph = UndirectedWeightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
         graph.addVertex(vertex1)
         graph.addVertex(vertex2)
         assertEquals(0, graph.addEdge(vertex1, vertex2, 1))
@@ -45,9 +45,9 @@ class GraphTests {
 
     @Test
     fun `undirected weighted graph get weight`() {
-        val graph = UndirectedWeightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
+        val graph = UndirectedWeightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
         graph.addVertex(vertex1)
         graph.addVertex(vertex2)
         graph.addEdge(vertex1, vertex2, 1)
@@ -56,9 +56,9 @@ class GraphTests {
 
     @Test
     fun `directed weighted graph add edge`() {
-        val graph = DirectedWeightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
+        val graph = DirectedWeightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
         graph.addVertex(vertex1)
         graph.addVertex(vertex2)
         assertEquals(0, graph.addEdge(vertex1, vertex2, 1))
@@ -66,10 +66,10 @@ class GraphTests {
 
     @Test
     fun `directed weighted graph getNeighbors correct`() {
-        val graph = DirectedWeightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
-        val vertex3 = Vertex(3, "3")
+        val graph = DirectedWeightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
+        val vertex3 = Vertex(3L, "3")
         graph.addVertex(vertex1)
         graph.addVertex(vertex2)
         graph.addVertex(vertex3)
@@ -80,15 +80,15 @@ class GraphTests {
 
     @Test
     fun `directed weighted graph getNeighbors with unknown vertex`() {
-        val graph = DirectedWeightedGraph<Int>()
+        val graph = DirectedWeightedGraph<Long>()
         assertEquals(emptyList(), graph.getNeighbors(Vertex(1, "1")))
     }
 
     @Test
     fun `directed weighted graph getNeighbors with edge direction is in wrong direction`() {
-        val graph = DirectedWeightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
+        val graph = DirectedWeightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
         graph.addVertex(vertex1)
         graph.addVertex(vertex2)
         graph.addEdge(vertex1, vertex2, 1)
@@ -97,9 +97,9 @@ class GraphTests {
 
     @Test
     fun `directed unweighted graph add edge`() {
-        val graph = DirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
+        val graph = DirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
         graph.addVertex(vertex1)
         graph.addVertex(vertex2)
         assertEquals(0, graph.addEdge(vertex1, vertex2))
@@ -107,10 +107,10 @@ class GraphTests {
 
     @Test
     fun `directed unweighted graph getNeighbors correct`() {
-        val graph = DirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
-        val vertex3 = Vertex(3, "3")
+        val graph = DirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
+        val vertex3 = Vertex(3L, "3")
         graph.addVertex(vertex1)
         graph.addVertex(vertex2)
         graph.addVertex(vertex3)
@@ -121,15 +121,15 @@ class GraphTests {
 
     @Test
     fun `directed unweighted graph getNeighbors with unknown vertex`() {
-        val graph = DirectedUnweightedGraph<Int>()
+        val graph = DirectedUnweightedGraph<Long>()
         assertEquals(emptyList(), graph.getNeighbors(Vertex(1, "1")))
     }
 
     @Test
     fun `directed unweighted graph getNeighbors with edge direction is in wrong direction`() {
-        val graph = DirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
+        val graph = DirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
         graph.addVertex(vertex1)
         graph.addVertex(vertex2)
         graph.addEdge(vertex1, vertex2)
@@ -138,28 +138,28 @@ class GraphTests {
 
     @Test
     fun `undirected unweighted graph getVertex`() {
-        val graph = UndirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
+        val graph = UndirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
         graph.addVertex(vertex1)
         assertEquals(vertex1, graph.getVertex(1))
     }
 
     @Test
     fun `undirected unweighted graph getEdge`() {
-        val graph = UndirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
+        val graph = UndirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
         graph.addVertex(vertex1)
         graph.addVertex(vertex2)
         graph.addEdge(vertex1, vertex2)
-        assertEquals(UnweightedEdge(0, vertex1, vertex2), graph.getEdge(0))
+        assertEquals(UnweightedEdge(0L, vertex1, vertex2), graph.getEdge(0))
     }
 
     @Test
     fun `undirected unweighted graph getAllVertex`() {
-        val graph = UndirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
+        val graph = UndirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
         graph.addVertex(vertex1)
         graph.addVertex(vertex2)
         assertEquals(listOf(vertex1, vertex2), graph.getAllVertices())
@@ -167,26 +167,26 @@ class GraphTests {
 
     @Test
     fun `undirected unweighted graph getAllEdge`() {
-        val graph = UndirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
-        val vertex3 = Vertex(3, "3")
-        val vertex4 = Vertex(4, "4")
+        val graph = UndirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
+        val vertex3 = Vertex(3L, "3")
+        val vertex4 = Vertex(4L, "4")
         graph.addVertex(vertex1)
         graph.addVertex(vertex2)
         graph.addVertex(vertex3)
         graph.addVertex(vertex4)
         graph.addEdge(vertex1, vertex2)
         graph.addEdge(vertex3, vertex4)
-        assertEquals(listOf(UnweightedEdge(0, vertex1, vertex2), UnweightedEdge(1, vertex3, vertex4)), graph.getAllEdges())
+        assertEquals(listOf(UnweightedEdge(0L, vertex1, vertex2), UnweightedEdge(1L, vertex3, vertex4)), graph.getAllEdges())
     }
 
     @Test
     fun `undirected unweighted graph getNeighbors correct version 1`() {
-        val graph = UndirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
-        val vertex3 = Vertex(3, "3")
+        val graph = UndirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
+        val vertex3 = Vertex(3L, "3")
         graph.addVertex(vertex1)
         graph.addVertex(vertex2)
         graph.addVertex(vertex3)
@@ -197,10 +197,10 @@ class GraphTests {
 
     @Test
     fun `undirected unweighted graph getNeighbors correct version 2`() {
-        val graph = UndirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
-        val vertex3 = Vertex(3, "3")
+        val graph = UndirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
+        val vertex3 = Vertex(3L, "3")
         graph.addVertex(vertex1)
         graph.addVertex(vertex2)
         graph.addVertex(vertex3)
@@ -211,33 +211,33 @@ class GraphTests {
 
     @Test
     fun `undirected unweighted graph getNeighbors with unknown vertex`() {
-        val graph = UndirectedUnweightedGraph<Int>()
+        val graph = UndirectedUnweightedGraph<Long>()
         assertEquals(emptyList(), graph.getNeighbors(Vertex(1, "1")))
     }
 
     @Test
     fun `undirected unweighted graph addEdge with unknown from vertex`() {
-        val graph = UndirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
+        val graph = UndirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
         graph.addVertex(vertex1)
         assertEquals(null, graph.addEdge(vertex2, vertex1))
     }
 
     @Test
     fun `undirected unweighted graph addEdge with unknown to vertex`() {
-        val graph = UndirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
+        val graph = UndirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
         graph.addVertex(vertex1)
         assertEquals(null, graph.addEdge(vertex1, vertex2))
     }
 
     @Test
     fun `undirected unweighted graph twice addEdge`() {
-        val graph = UndirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
+        val graph = UndirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
         graph.addVertex(vertex1)
         graph.addVertex(vertex2)
         graph.addEdge(vertex1, vertex2)
@@ -246,9 +246,9 @@ class GraphTests {
 
     @Test
     fun `undirected unweighted graph clear`() {
-        val graph = UndirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
+        val graph = UndirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
         assertEquals(true, graph.vertices.isEmpty())
         graph.addVertex(vertex1)
         assertEquals(false, graph.vertices.isEmpty())
@@ -263,26 +263,26 @@ class GraphTests {
 
     @Test
     fun `undirected unweighted graph removeVertex correct`() {
-        val graph = UndirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
+        val graph = UndirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
         graph.addVertex(vertex1)
         assertEquals(1, graph.removeVertex(1))
     }
 
     @Test
     fun `undirected unweighted graph removeVertex with unknown id`() {
-        val graph = UndirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
+        val graph = UndirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
         graph.addVertex(vertex1)
         assertEquals(null, graph.removeVertex(52))
     }
 
     @Test
     fun `undirected unweighted graph removeVertex with edges version 1`() {
-        val graph = UndirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(3, "2")
-        val vertex3 = Vertex(3, "3")
+        val graph = UndirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(3L, "2")
+        val vertex3 = Vertex(3L, "3")
         graph.addVertex(vertex1)
         graph.addVertex(vertex2)
         graph.addVertex(vertex3)
@@ -293,10 +293,10 @@ class GraphTests {
 
     @Test
     fun `undirected unweighted graph removeVertex with edges version 2`() {
-        val graph = UndirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(3, "2")
-        val vertex3 = Vertex(3, "3")
+        val graph = UndirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(3L, "2")
+        val vertex3 = Vertex(3L, "3")
         graph.addVertex(vertex1)
         graph.addVertex(vertex2)
         graph.addVertex(vertex3)
@@ -307,10 +307,10 @@ class GraphTests {
 
     @Test
     fun `undirected unweighted graph removeVertex with edges version 3`() {
-        val graph = UndirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
-        val vertex3 = Vertex(3, "3")
+        val graph = UndirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
+        val vertex3 = Vertex(3L, "3")
         graph.addVertex(vertex1)
         graph.addVertex(vertex2)
         graph.addVertex(vertex3)
@@ -320,9 +320,9 @@ class GraphTests {
 
     @Test
     fun `undirected unweighted graph removeEdge by id`() {
-        val graph = UndirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
+        val graph = UndirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
         graph.addVertex(vertex1)
         graph.addVertex(vertex2)
         graph.addEdge(vertex1, vertex2)
@@ -331,9 +331,9 @@ class GraphTests {
 
     @Test
     fun `undirected unweighted graph removeEdge by unknown id`() {
-        val graph = UndirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
+        val graph = UndirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
         graph.addVertex(vertex1)
         graph.addVertex(vertex2)
         graph.addEdge(vertex1, vertex2)
@@ -342,9 +342,9 @@ class GraphTests {
 
     @Test
     fun `undirected unweighted graph removeEdge by from and to`() {
-        val graph = UndirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
+        val graph = UndirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
         graph.addVertex(vertex1)
         graph.addVertex(vertex2)
         graph.addEdge(vertex1, vertex2)
@@ -353,9 +353,9 @@ class GraphTests {
 
     @Test
     fun `undirected unweighted graph removeEdge by from and to with wrong direction`() {
-        val graph = UndirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
+        val graph = UndirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
         graph.addVertex(vertex1)
         graph.addVertex(vertex2)
         graph.addEdge(vertex2, vertex1)
@@ -364,9 +364,9 @@ class GraphTests {
 
     @Test
     fun `undirected unweighted graph removeEdge by from and to with unknown from`() {
-        val graph = UndirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
+        val graph = UndirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
         graph.addVertex(vertex1)
         graph.addVertex(vertex2)
         graph.addEdge(vertex1, vertex2)
@@ -375,9 +375,9 @@ class GraphTests {
 
     @Test
     fun `undirected unweighted graph removeEdge by from and to with unknown to`() {
-        val graph = UndirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
+        val graph = UndirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
         graph.addVertex(vertex1)
         graph.addVertex(vertex2)
         graph.addEdge(vertex1, vertex2)
@@ -386,9 +386,9 @@ class GraphTests {
 
     @Test
     fun `undirected unweighted graph removeEdge by unknown from and unknown to`() {
-        val graph = UndirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
+        val graph = UndirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
         graph.addVertex(vertex1)
         graph.addVertex(vertex2)
         graph.addEdge(vertex1, vertex2)
@@ -397,10 +397,10 @@ class GraphTests {
 
     @Test
     fun `undirected unweighted graph removeEdge by from and to with wrong to`() {
-        val graph = UndirectedUnweightedGraph<Int>()
-        val vertex1 = Vertex(1, "1")
-        val vertex2 = Vertex(2, "2")
-        val vertex3 = Vertex(3, "3")
+        val graph = UndirectedUnweightedGraph<Long>()
+        val vertex1 = Vertex(1L, "1")
+        val vertex2 = Vertex(2L, "2")
+        val vertex3 = Vertex(3L, "3")
         graph.addVertex(vertex1)
         graph.addVertex(vertex2)
         graph.addVertex(vertex3)
